pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "QRate"
version = "0.1.0"
description = "A QC data curation tool"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Himal Shrestha", email = "himal.shrestha@unimelb.edu.au"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
requires-python = ">=3.8"
dependencies = [
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov",
    "black",
    "flake8",
]

[project.scripts]
qrate = "qrate.main:main"

[project.urls]
Homepage = "https://github.com/MDU-PHL/QCheck"
Repository = "https://github.com/MDU-PHL/QCheck"
Issues = "https://github.com/MDU-PHL/QCheck/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["qrate*"]

[tool.setuptools.package-data]
qrate = ["config/*.yaml"]
